Save new command
